engine-core: port FUN_801D4A60 - four cutscene programs, not one approach controller
`801d4a60` dispatches on `+0x54` through a 38-word table at `0x801CE960`, and
the entry state computes its own successor as
`+0x54 = (+0x54 + 1) + (+0x50) * 10`. So `+0x50` is a PROGRAM SELECTOR and the
state space is four programs on a stride of ten, entered at 1 / 11 / 21 / 31 -
exactly the table's live blocks. The fifteen slots pointing at the epilogue
(6..10, 16..20, 27..30) are the unused tails of each ten-wide block, not
fifteen dead states. Without the x10 the table reads as a sparse mess; with it,
four short programs of 5, 5, 6 and 7 states.

Programs 0 and 1 are openers, 2 and 3 their closers. The MAN loader
FUN_8003AEB0 spawns program 2 when flag 0x17 is set (0x8003BB10) and program 3
when flag 0x0C is set (0x8003BB38) - precisely the flags 0 and 1 SET and 2 and
3 CLEAR. So the loader is not starting cutscenes, it is finishing ones a scene
change interrupted. That rule is ported and live off `load_scene`, so the actor
is spawned in a real session; the step itself is disclosed, with the missing
input named exactly (the BGM request/ack pair and the CD-XA in-flight counter
have no engine counterparts, and three states park indefinitely on them).

What the programs do: stage move-VM parts at the player once per vsync, lock
the player, swap BGM to 0x7F3, and play XA17. The audio census had already
reached this function from the other side - `audio.md` lists field 0897
0x801D4FCC as clip 0x10, "scripted-scene voice stream". That is program 2's
state 0x16.

Also ported: FUN_801D5A24 (the spawner; its `value` argument is the program id).

Provenance: `overlay_0897_801d4a60.txt` stops at 690 instructions. Five
live-RAM field captures say 756, and so does capstone over the static image at
the committed base (file 0x006248 -> `jr ra` at 0x801D5628). The 66 dropped
instructions are states 0x22..0x25 plus the shared tail - most of program 3 -
which is how a four-program machine came to be documented as one controller.

+++ crates/engine-core/src/world/frame_tick.rs
@@ -367,7 +367,7 @@ impl World {
             // any marked by the scene-transition sweep or a `FUN_8003CF40`
             // retire since the last tick.
             // REF: FUN_8002519C
-            self.tick_handler_actors(cadence as u8);
+            self.tick_handler_actors(cadence);
             self.tick_actors();
             // Actor-VM glides (op 0x09 `MotionAt` -> `start_motion`): one
             // motion-VM pursue step per game tick toward the recorded target.
